Add ModalDrawerSheet example
Added a new screen named `Modal Drawer Sheet` that demonstrates the use of `ModalDrawerSheet`.
This screen allows users to open a drawer sheet from the right or left side of the screen using chevrons.
The drawer sheet contains a simple list item and a top app bar with a back button.