Configure alpha build type and use it in CI
This commit introduces an `alpha` build type with a corresponding `applicationIdSuffix` and `versionNameSuffix`.
The CI workflow is updated to use the `alpha` build type, including assemble, test, lint, and detekt tasks.
Also, the `config` property for `detekt` task is changed to use `setFrom()` to properly override the default config.
The dummy signing value configuration is removed from `build.gradle.kts`.