Refactor: Introduce `AppModalDrawerSheet` and simplify `ModalDrawerSheetScreen`
This commit introduces a new composable, `AppModalDrawerSheet`, which encapsulates the logic for displaying a modal drawer sheet.
It handles drawer state, layout direction, and tap gestures for closing the drawer.

The `ModalDrawerSheetScreen` has been refactored to use `AppModalDrawerSheet` for a cleaner and more maintainable implementation.
This improves code readability and reusability.
The `ModalBottomSheetScreen` was also updated to display a button to enable/disable the top padding.